Prevent deletion of special folders in e-mail like Sent items, Trash and Drafts

+++ www/modules/email/controller/FolderController.php
@@ -53,6 +53,9 @@ class GO_Email_Controller_Folder extends GO_Base_Controller_AbstractController {
 		$account = GO_Email_Model_Account::model()->findByPk($params['account_id']);
 				
 		$mailbox = new GO_Email_Model_ImapMailbox($account, array("name"=>$params["mailbox"]));
+		if($mailbox->isSpecial())
+			throw new Exception(GO::t("cantDeleteSpecialFolder","email"));
+		
 		$success = $mailbox->delete();
 		
 		return array("success"=>$success);
@@ -64,15 +67,15 @@ class GO_Email_Controller_Folder extends GO_Base_Controller_AbstractController {
 		$mailbox = new GO_Email_Model_ImapMailbox($account, array("name"=>$params["mailbox"]));
                 
                 
-                if(!empty($account->trash) && $params["mailbox"] != $account->trash) {
-                    $imap = $account->openImapConnection($params["mailbox"]);
-                    $uids = $imap->sort_mailbox();
-                    $imap->set_message_flag($uids, "\Seen");
-                    $success=$imap->move($uids,$account->trash);
-                }else {
-                    $success = $mailbox->truncate();
-                }
-                    
+		if(!empty($account->trash) && $params["mailbox"] != $account->trash) {
+				$imap = $account->openImapConnection($params["mailbox"]);
+				$uids = $imap->sort_mailbox();
+				$imap->set_message_flag($uids, "\Seen");
+				$success=$imap->move($uids,$account->trash);
+		}else {
+				$success = $mailbox->truncate();
+		}
+
 		
 		return array("success"=>$success);
 	}
